Working sequential approach

+++ src/algorithms/column_generation.h
@@ -24,7 +24,7 @@ struct CGResult {
     std::vector<double> integer_solution_costs;
     std::vector<double> covered_interventions;
     std::vector<int> integer_covered_interventions;
-    std::vector<int> time_points;
+    std::vector<double> time_points;
 };
